fixed sign comparison warnings

+++ src/Mixer/broyden1_mixer.hpp
@@ -82,7 +82,7 @@ class Broyden1 : public Mixer<FUNCS...>
             S.zero();
             mdarray<double, 2> S_local(history_size, history_size);
             S_local.zero();
-            for (int j1 = 0; j1 < history_size; j1++) {
+            for (int j1 = 0; j1 < static_cast<int>(history_size); j1++) {
                 int i1 = this->idx_hist(this->step_ - j1);
                 int i2 = this->idx_hist(this->step_ - j1 - 1);
                 this->copy(this->residual_history_[i1], this->tmp1_);
@@ -98,8 +98,8 @@ class Broyden1 : public Mixer<FUNCS...>
                 }
             }
             this->comm_.allreduce(S.at(memory_t::host), (int)S.size());
-            for (int j1 = 0; j1 < history_size; j1++) {
-                for (int j2 = 0; j2 < history_size; j2++) {
+            for (int j1 = 0; j1 < static_cast<int>(history_size); j1++) {
+                for (int j2 = 0; j2 < static_cast<int>(history_size); j2++) {
                     S(j1, j2) += S_local(j1, j2);
                 }
             }
@@ -107,7 +107,7 @@ class Broyden1 : public Mixer<FUNCS...>
             /* invert matrix */
             linalg<device_t::CPU>::syinv(history_size, S);
             /* restore lower triangular part */
-            for (int j1 = 0; j1 < history_size; j1++) {
+            for (int j1 = 0; j1 < static_cast<int>(history_size); j1++) {
                 for (int j2 = 0; j2 < j1; j2++) {
                     S(j1, j2) = S(j2, j1);
                 }
@@ -117,7 +117,7 @@ class Broyden1 : public Mixer<FUNCS...>
             c.zero();
             mdarray<double, 1> c_local(history_size);
             c_local.zero();
-            for (int j = 0; j < history_size; j++) {
+            for (int j = 0; j < static_cast<int>(history_size); j++) {
                 int i1 = this->idx_hist(this->step_ - j);
                 int i2 = this->idx_hist(this->step_ - j - 1);
 
@@ -128,13 +128,13 @@ class Broyden1 : public Mixer<FUNCS...>
                 c_local(j) = this->inner_product(true, this->tmp1_, this->residual_history_[idx_step]);
             }
             this->comm_.allreduce(c.at(memory_t::host), (int)c.size());
-            for (int j = 0; j < history_size; j++) {
+            for (int j = 0; j < static_cast<int>(history_size); j++) {
                 c(j) += c_local(j);
             }
 
-            for (int j = 0; j < history_size; j++) {
+            for (int j = 0; j < static_cast<int>(history_size); j++) {
                 double gamma = 0;
-                for (int i = 0; i < history_size; i++) {
+                for (int i = 0; i < static_cast<int>(history_size); i++) {
                     gamma += c(i) * S(i, j);
                 }
